Fix notification updates
- Before, the entire browser refreshed when receiving a new
notification. This was because the entire App would show a circular
progress when the getNotification() request was tracked. Upon removing
this loading state from the component, the browser no longer the entire
dashboard when receiving new notifications. If a user has dialogs open,
those dialogs will remain open instead of being closed.